changing tsdb format for csv 2 tsdb
