+++ httpblaster/igz_data/igz_tsdb_item.go
@@ -3,34 +3,54 @@ package igz_data
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/v3io/v3io-tsdb/pkg/utils"
 	"strconv"
 )
 
-type IgzTSDBItem struct {
-	Lset utils.Labels
-	Time string
-	Value float64
+
+type Sample struct{
+	T string `json:"t"`
+	V map[string]float64 `json:"v"`
 }
 
-type IgzTSDBItems struct {
+type IgzTSDBItem struct {
+	Metric string	`json:"metric"`
+
+	Labels  map[string]string `json:"labels"`
+	Samples []Sample          `json:"samples"`
+}
+
+func (self *IgzTSDBItem) GenerateStruct(vals []string,parser *EmdSchemaParser) error{
+	self.InsertParserMetric(vals,parser)
+	self.InsertParserLables(vals,parser)
+	self.InsertParserSample(vals ,parser)
+	return nil
+}
+
+type IgzTSDBItems2 struct {
 	Items []IgzTSDBItem
 }
 
-
-func (self *IgzTSDBItem) GenerateStructs(vals []string,parser *EmdSchemaParser) ([]IgzTSDBItem,error){
-	self.InsertTSDBName(vals,parser)
-	self.InsertTime(vals ,parser)
-	self.InsertValue(vals[parser.tsdb_value_index])
-	return nil,nil
+func (self *IgzTSDBItem) InsertMetric(metric string){
+	self.Metric = metric
 }
 
 
-func (self *IgzTSDBItem) GenerateStruct(vals []string,parser *EmdSchemaParser) error{
-	self.InsertTSDBName(vals,parser)
-	self.InsertTime(vals ,parser)
-	self.InsertValue(vals[parser.tsdb_value_index])
-	return nil
+func (self *IgzTSDBItem) InsertLable(key string,value string){
+	if len(self.Labels) == 0 {
+		self.Labels = make( map[string]string)
+	}
+	self.Labels[key] = value
+}
+
+func (self *IgzTSDBItem) InsertLables(lables map[string]string){
+	self.Labels =lables
+}
+
+func (self *IgzTSDBItem) InsertSample(ts string,value float64){
+	s:= &Sample{}
+	s.T =ts
+	s.V = map[string]float64{"n":value}
+	self.Samples = append(self.Samples,*s)
 }
 
 
@@ -39,48 +59,37 @@ func (self *IgzTSDBItem) ToJsonString() string {
 	return string(body)
 }
 
-func (self *IgzTSDBItem) InsertTSDBName(vals []string,parser *EmdSchemaParser) error {
+
+func (self *IgzTSDBItem) InsertParserMetric(vals []string,parser *EmdSchemaParser)  {
 	parser.tsdb_name_index=GetIndexByValue(parser.values_map,parser.tsdb_name)
 	input :=""
 	if parser.tsdb_name_index > -1 {
 		input = vals[parser.tsdb_name_index]
 	}	else {
-	input = parser.tsdb_name
+		input = parser.tsdb_name
 	}
-	self.InsertName(input)
-	for key, val := range parser.tsdb_attributes_map {
-		lable := utils.Label{Name: key, Value: vals[val]}
-		self.Lset=  append(self.Lset,lable)
-		}
-	return nil
+	self.InsertMetric(input)
 }
 
-func (self *IgzTSDBItem) InsertTime(vals []string,parser *EmdSchemaParser) error {
+func (self *IgzTSDBItem) InsertParserLables(vals []string,parser *EmdSchemaParser) {
+	for key, val := range parser.tsdb_attributes_map {
+		self.InsertLable(key,vals[val])
+	}
+}
+
+func (self *IgzTSDBItem) InsertParserSample(vals []string,parser *EmdSchemaParser) {
 	for _, v := range parser.values_map {
 		if v.Name == parser.tsdb_time {
 			parser.tsdb_time_index = v.Index
 		}
 	}
-	input := vals[parser.tsdb_time_index]
-	//add validation on time
-	self.InsertTimeString(input)
-	return nil
-}
-
-func (self *IgzTSDBItem) InsertTimeString(strVal string){
-	self.Time=strVal
-}
-
-func (self *IgzTSDBItem) InsertName(strVal string){
-	self.Lset = utils.Labels{{Name: "__name__",Value:strVal}}
-}
-
-func (self *IgzTSDBItem) InsertValue(strVal string){
-	f, err := strconv.ParseFloat(strVal, 64)
+	ts := vals[parser.tsdb_time_index]
+	val := vals[parser.tsdb_value_index]
+	f, err := strconv.ParseFloat(val, 64)
 	if err!=nil {
-		panic(fmt.Sprintf("conversion error to float %v ", strVal))
+		panic(fmt.Sprintf("conversion error to float %v ", val))
 	}
-	self.Value=f
+	self.InsertSample(ts,f)
 }
 
 func GetIndexByValue(vals map[int]SchemaValue,val string) (int){
